Move functions to separate module, fix setting of profile value
