+++ monitor.py
@@ -1,13 +1,13 @@
-import time
 import os
-import securedrop
-import json
+import time
 from typing import Optional, Union, Dict
 
 import requests
 from boto3 import Session
-from botocore.exceptions import ClientError
 from requests.exceptions import RequestException
+
+import securedrop
+from notifications import create_message, send_message, send_email
 
 CHARSET = "UTF-8"
 
@@ -34,96 +34,6 @@ def fetch_parameter(client, name: str) -> Union[str, None]:
     except client.exceptions.ParameterNotFound:
         print(f'parameter not found: {name}')
         return None
-
-
-def generate_text(heading: str, text: str) -> str:
-    return f"""
-    {heading}\r\n
-    {text}\n
-    This email was sent by the Secure Contact application
-    """
-
-
-def generate_html(heading: str, text: str) -> str:
-    return f"""<html>
-    <head></head>
-    <body>
-    <h1>{heading}</h1>
-    <p>{text}</p>
-    <p>This email was sent by the Secure Contact application using <a href='https://aws.amazon.com/ses/'>AWS SES</a></p>
-    </body>
-    </html>"""
-
-
-def create_message(heading: str, text: str):
-    body_html = generate_html(heading, text)
-    body_text = generate_text(heading, text)
-    return {
-        'Body': {
-            'Html': {
-                'Charset': CHARSET,
-                'Data': body_html,
-            },
-            'Text': {
-                'Charset': CHARSET,
-                'Data': body_text,
-            },
-        },
-        'Subject': {
-            'Charset': CHARSET,
-            'Data': '[ALERT P1] SecureDrop Site Failing Healthcheck',
-        },
-    }
-
-
-def send_email(session: Session, config: Dict[str, str], message: Dict):
-    ses_client = session.client('ses')
-    sender = config['PRODMON_SENDER']
-    recipient = config['PRODMON_RECIPIENT']
-    try:
-        response = ses_client.send_email(
-            Destination={
-                'ToAddresses': [
-                    recipient,
-                ],
-            },
-            Message=message,
-            Source=f'SecureDrop Monitor <{sender}>',
-        )
-    # TODO: use logging library instead and send logs somewhere sensible
-    except ClientError as e:
-        print(e.response['Error']['Message'])
-    else:
-        print('Email sent! Message ID:')
-        print(response['MessageId'])
-
-
-def generate_card(title: str, subtitle: str) -> Dict:
-    return {
-        "cards": [
-            {
-                "header": {
-                    "title": title,
-                    "subtitle": subtitle
-                }
-            }
-        ]
-    }
-
-
-def send_message(config: Dict[str, str], passed: bool):
-    # TODO: message @all to notify when healthcheck fails
-    headers = {'Content-Type': 'application/json; charset=UTF-8'}
-
-    status = 'Status: 💚💚💚' if passed else 'Status: 💔💔💔'
-    card = json.dumps(generate_card('SecureDrop Monitor', status))
-
-    try:
-        response = requests.post(url=config['PRODMON_WEBHOOK'], headers=headers, data=card)
-        print(f'Message sent to Hangouts Chat!')
-        print(f'Status code {response.status_code} returned from chat.googleapis.com')
-    except RequestException as err:
-        print(err)
 
 
 # N.B. this script requires Tor to be running on the server
@@ -199,7 +109,7 @@ def run(session: Session, config: Dict[str, str]):
 
 if __name__ == '__main__':
     STAGE = get_stage('/etc/stage')
-    AWS_PROFILE = os.getenv('AWS_PROFILE') if os.getenv('AWS_PROFILE') else None
+    AWS_PROFILE = 'infosec' if STAGE == 'DEV' else None
     SESSION = create_session(profile=AWS_PROFILE)
     SSM_CLIENT = SESSION.client('ssm')
